apps/backend/chatbot-service/requirements.txt
fastapi
uvicorn[standard]
pydantic

mysqlclient
SQLAlchemy

yfinance

cachetools

google-genai